+++ derive_builder/CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - customize setter names via `#[builder(setter(name="..."))]`
+- Implement `TryFrom` from builder to target type from generated builder 
+  when `try_from` crate feature is enabled.
 
 ## [0.4.4] - 2017-04-12